Add 'lowCodeCacheFreeSpace' field to J9JITConfig
This field is set to 1 by the JIT when it detects that
that no additional code cache can be allocated and all
of the allocated code caches are very low on free space.
This information can be used by other VM components. For
example, the GC can use it to trigger class unloading
more aggressively, to make more room in the code caches.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>